# openapi.json
{"openapi":"3.1.0","info":{"title":"Shopping Assistant Backend","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/chat":{"post":{"tags":["chat"],"summary":"Chat","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/device/token":{"post":{"tags":["device"],"summary":"Upsert Device Token","operationId":"upsert_device_token_device_token_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertDeviceTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/watch":{"post":{"tags":["watch"],"summary":"Add Watch","operationId":"add_watch_watch_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["watch"],"summary":"List Watch","operationId":"list_watch_watch_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/discover/search":{"get":{"tags":["discover"],"summary":"Search","operationId":"search_discover_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiscoverSearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assist/checkout-plan":{"post":{"tags":["assist"],"summary":"Checkout Plan","operationId":"checkout_plan_assist_checkout_plan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutPlanRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutPlanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/watch/refresh":{"post":{"tags":["watch"],"summary":"Refresh Prices","operationId":"refresh_prices_watch_refresh_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notifications":{"get":{"tags":["notifications"],"summary":"List Notifications","operationId":"list_notifications_notifications_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notifications/ack":{"post":{"tags":["notifications"],"summary":"Ack Notifications","operationId":"ack_notifications_notifications_ack_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agent/chat":{"post":{"tags":["agent"],"summary":"Agent Chat","operationId":"agent_chat_agent_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"message":{"type":"string","title":"Message"},"conversation_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Conversation Id"}},"type":"object","required":["user_id","message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"conversation_id":{"type":"integer","title":"Conversation Id"},"reply":{"type":"string","title":"Reply"}},"type":"object","required":["conversation_id","reply"],"title":"ChatResponse"},"CheckoutPlanRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"url":{"type":"string","title":"Url"},"product_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Hint"}},"type":"object","required":["user_id","url"],"title":"CheckoutPlanRequest"},"CheckoutPlanResponse":{"properties":{"retailer_domain":{"type":"string","title":"Retailer Domain"},"steps":{"items":{"type":"string"},"type":"array","title":"Steps"},"notes":{"items":{"type":"string"},"type":"array","title":"Notes","default":[]}},"type":"object","required":["retailer_domain","steps"],"title":"CheckoutPlanResponse"},"DiscoverResult":{"properties":{"title":{"type":"string","title":"Title"},"url":{"type":"string","title":"Url"},"snippet":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Snippet"},"source":{"type":"string","title":"Source","default":"serpapi"},"retailer_domain":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retailer Domain"}},"type":"object","required":["title","url"],"title":"DiscoverResult"},"DiscoverSearchResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/DiscoverResult"},"type":"array","title":"Results"}},"type":"object","required":["results"],"title":"DiscoverSearchResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UpsertDeviceTokenRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"platform":{"type":"string","title":"Platform","default":"ios"},"fcm_token":{"type":"string","title":"Fcm Token"}},"type":"object","required":["user_id","fcm_token"],"title":"UpsertDeviceTokenRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WatchCreateRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"url":{"type":"string","title":"Url"},"title_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title Hint"},"upc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Upc"},"product_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Key"},"desired_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Desired Price"},"currency":{"type":"string","title":"Currency","default":"USD"}},"type":"object","required":["user_id","url"],"title":"WatchCreateRequest"},"WatchItemOut":{"properties":{"url":{"type":"string","title":"Url"},"title_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title Hint"},"upc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Upc"},"product_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Key"},"desired_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Desired Price"},"currency":{"type":"string","title":"Currency","default":"USD"},"last_seen_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Last Seen Price"},"best_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Best Price"},"best_retailer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Retailer"},"best_offer_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Offer Url"},"last_checked_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Checked At"},"best_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Title"},"best_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Description"},"best_rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Best Rating"},"best_reviews_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Best Reviews Count"},"best_condition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Condition"},"best_seller_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Seller Type"}},"type":"object","required":["url"],"title":"WatchItemOut"},"WatchListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/WatchItemOut"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"WatchListResponse"}}}}